+++ zebra-rpc/src/methods/get_block_template_rpcs/types.rs
@@ -1,0 +1,6 @@
+//! Types used in mining RPC methods.
+
+pub(crate) mod coinbase;
+pub(crate) mod default_roots;
+pub(crate) mod get_block_template;
+pub(crate) mod transaction;
