+++ zebra-rpc/src/methods/get_block_template_rpcs/types/coinbase.rs
@@ -1,0 +1,5 @@
+//! The `Coinbase` type is part of the `getblocktemplate` RPC method output.
+
+/// documentation and fields to be added in #5453.
+#[derive(Clone, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize)]
+pub struct Coinbase {}
