+++ zebra-rpc/src/methods/get_block_template_rpcs/types/transaction.rs
@@ -1,0 +1,5 @@
+//! The `Transaction` type is part of the `getblocktemplate` RPC method output.
+
+/// Documentation and fields to be added in #5454.
+#[derive(Clone, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize)]
+pub struct Transaction {}
