+++ zebra-rpc/src/methods/tests/snapshot.rs
@@ -302,7 +302,7 @@ fn snapshot_rpc_getblockhash(block_hash: GetBlockHash, settings: &insta::Setting
 #[cfg(feature = "getblocktemplate-rpcs")]
 /// Snapshot `getblocktemplate` response, using `cargo insta` and JSON serialization.
 fn snapshot_rpc_getblocktemplate(
-    block_template: crate::methods::get_block_template::GetBlockTemplate,
+    block_template: crate::methods::get_block_template_rpcs::types::get_block_template::GetBlockTemplate,
     settings: &insta::Settings,
 ) {
     settings.bind(|| insta::assert_json_snapshot!("get_block_template", block_template));
